Progress monitoring for aleap

+++ src/main/java/com/metaverse/workflow/trainingandnontrainingtarget/dtos/TrainingTargetsAndAchievementsResponse.java
@@ -6,10 +6,10 @@ import lombok.Data;
 public class TrainingTargetsAndAchievementsResponse {
     private String activityName;
     private String financialYear;
-    private Integer trainingTargetQ1; // Physical Training target for Q1
-    private Integer trainingTargetQ2;
-    private Integer trainingTargetQ3;
-    private Integer trainingTargetQ4;
+    private Long trainingTargetQ1; // Physical Training target for Q1
+    private Long trainingTargetQ2;
+    private Long trainingTargetQ3;
+    private Long trainingTargetQ4;
     private Double financialTargetQ1;
     private Double financialTargetQ2;
     private Double financialTargetQ3;
